Pylint fixes for develop

+++ salt/cloud/clouds/vmware.py
@@ -44,9 +44,10 @@ cloud configuration at
       password: verybadpass
       url: vcenter03.domain.com
 '''
-from __future__ import absolute_import
 
 # Import python libs
+from __future__ import absolute_import
+from random import randint
 import pprint
 import logging
 import time
@@ -236,10 +237,9 @@ def _edit_existing_hard_disk_helper(disk, size_kb):
 
 
 def _add_new_hard_disk_helper(disk_label, size_gb, unit_number):
-    from random import randint
     random_key = randint(-2099, -2000)
 
-    size_kb = long(size_gb) * 1024 * 1024
+    size_kb = int(size_gb) * 1024 * 1024
 
     disk_spec = vim.vm.device.VirtualDeviceSpec()
     disk_spec.fileOperation = 'create'
@@ -274,7 +274,6 @@ def _edit_existing_network_adapter_helper(network_adapter, new_network_name):
 
 
 def _add_new_network_adapter_helper(network_adapter_label, network_name, adapter_type):
-    from random import randint
     random_key = randint(-4099, -4000)
 
     adapter_type.strip().lower()
@@ -332,7 +331,7 @@ def _manage_devices(devices, vm):
                 existing_disks_label.append(device.deviceInfo.label)
                 if device.deviceInfo.label in devices['disk'].keys():
                     size_gb = devices['disk'][device.deviceInfo.label]['size']
-                    size_kb = long(size_gb) * 1024 * 1024
+                    size_kb = int(size_gb) * 1024 * 1024
                     if device.capacityInKB < size_kb:
                         # expand the disk
                         disk_spec = _edit_existing_hard_disk_helper(device, size_kb)
